Added markup API for tileboard activation, smoothify-movement, pingpong-mover.
This makes it so tileboards can be created and used without needing to
write any JS. The auto-activation of table.board elements can be disabled
by adding a data-no-auto-activate attribute to any element, though perhaps
we should copy bootstrap's way of doing that.

Just add the 'smoothify-movement' class to any table.board to auto-activate
the smoothify-movement plugin on it. And add the 'pingpong-mover' class to
any piece to automatically make it ping-pong once it starts moving (that
can be done through the markup API by setting the 'js-moving' class on a
piece and setting its 'data-moving-from' attribute to a tile ID).

Need to start documenting these APIs soon!